requirements.txt
streamlit
numpy
matplotlib
pillow
pymongo[srv]
gridfs